requirements.txt
asttokens==2.4.0
backcall==0.2.0
backports.functools-lru-cache==1.6.5
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
comm==0.1.4
contourpy==1.0.5
cryptography==41.0.3
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
filelock==3.9.0
fonttools==4.25.0
gmpy2==2.1.2
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.3.1
jupyter_core==4.12.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
munkres==1.1.4
nest-asyncio==1.5.6
networkx==3.1
numexpr==2.8.4
numpy==1.25.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.2.1
ply==3.11
prompt-toolkit==3.0.39
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-metric-learning==1.6.3
pytz==2022.7
pyzmq==25.1.0
requests==2.31.0
scikit-learn==1.2.2
scipy==1.11.1
setuptools==68.0.0
sip==6.6.2
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
threadpoolctl==2.2.0
toml==0.10.2
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.10.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
wheel==0.38.4
zipp==3.16.2
